+++ clHCA.h
@@ -249,12 +249,12 @@ PACKED(
 
 	struct stChannel
 	{
-		float wave[8][0x80];
+		double wave[8][0x80];
 		float base[0x80];
 		float block[0x80];
 		float wav1[0x80];
 		float wav2[0x80];
-		float wav3[0x80];
+		double wav3[0x80];
 		uint8_t value[0x80];
 		uint8_t scale[0x80];
 		uint8_t value2[8];
